// frontend/penguin/src/services/semanticGeneration/types.ts
export interface ObjectDescription {
  description: string;
  location: string;
  relationship: string;
  relative_size: string;
  shape_and_color: string;
  texture: string;
  appearance_details: string;
  orientation: string;
  number_of_objects?: number;
  pose?: string;
  expression?: string;
  action?: string;
}

export interface LightingDescription {
  conditions: string;
  direction: string;
  shadows: string;
}

export interface AestheticsDescription {
  composition: string;
  color_scheme: string;
  mood_atmosphere: string;
  preference_score: string;
  aesthetic_score: string;
}

export interface PhotographicDescription {
  depth_of_field: string;
  focus: string;
  camera_angle: string;
  lens_focal_length: string;
}

export interface TextRender {
  text: string;
  position: string;
  style: string;
}

export interface SemanticJSON {
  short_description: string;
  objects: ObjectDescription[];
  background_setting: string;
  lighting: LightingDescription;
  aesthetics: AestheticsDescription;
  photographic_characteristics: PhotographicDescription;
  style_medium: string;
  context?: string;
  artistic_style?: string;
  text_render?: TextRender[];
}